Support C++20

+++ Engine/json.hpp
@@ -261,7 +261,7 @@ namespace json {
 
         inline int length() const {
             if (Type == Class::Array)
-                return Internal.List->size();
+                return static_cast<int>(Internal.List->size());
             else
                 return -1;
         }
@@ -274,9 +274,9 @@ namespace json {
 
         inline int size() const {
             if (Type == Class::Object)
-                return Internal.Map->size();
+                return static_cast<int>(Internal.Map->size());
             else if (Type == Class::Array)
-                return Internal.List->size();
+                return static_cast<int>(Internal.List->size());
             else
                 return -1;
         }
